requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1
scipy==1.3.0
sewar==0.4.2
six==1.12.0
toml==0.10.0
torch==1.1.0
torchvision==0.3.0
